+++ modules/dashboard/templates/index.php
@@ -127,8 +127,8 @@ $upgrade_guide_url = admin_url( 'admin.php' ) . '?page=kmcf7-message-filter-opti
                     </div>
                     <h5> Here are a few changes in this version:</h5>
                     <ol>
-                        <li>Fix bug when adding a custom field on the WPForm settings page.</li>
-                        <li>Fix WPForm text fields not checked for spam.</li>
+                        <li>Fix bug with the CF7 Conditional Fields Plugin.</li>
+                        <li>Add Ability to select the forms to validate or not to validate (Blacklist/Whitelist).</li>
                     </ol>
 					<?php /*esc_html_e( "Please help translate this plugin to your language", KMCF7MS_TEXT_DOMAIN ) */?><!-- <a
                             href="https://translate.wordpress.org/projects/wp-plugins/cf7-message-filter/"
@@ -253,7 +253,7 @@ $words .= "]";
 
     var word_chart = new ApexCharts(document.querySelector("#words-chart"), word_options);
     word_chart.render();
-	<?php if(get_option( 'kmcfmf_version', '0' ) != KMCFMessageFilter::getVersion()):?>
+	<?php if(get_option( 'kmcfmf_version', '0' ) == KMCFMessageFilter::getVersion()):?>
     jQuery(document).ready(function ($) {
         $('#myModal').modal()
     });