+++ modules/settings/templates/contactform7.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace km_message_filter;
-$tags                  = ContactForm7Module::getTags();
+$tags                 = ContactForm7Module::getTags();
 $text_fields          = $tags['text'];
 $selected_text_fields = get_option( 'kmcfmf_tags_by_name' );
 $selected_text_fields = explode( ',', $selected_text_fields );
@@ -25,6 +25,8 @@ $selected_email_fields = array_map( function ( $e ) {
 	return array( 'text' => $e, 'value' => $e );
 }, $selected_email_fields );
 $email_fields          = array_merge( $email_fields, $selected_email_fields );
+
+$contact_forms = ContactForm7Module::getForms();
 ?>
     <h1><?php esc_html_e( "Contact Form 7 Settings ", KMCF7MS_TEXT_DOMAIN ) ?></h1>
 	<?php if ( ! is_plugin_active( 'contact-form-7/wp-contact-form-7.php' ) ): ?>
@@ -80,7 +82,23 @@ $email_fields          = array_merge( $email_fields, $selected_email_fields );
                     }
                 }
             });
+            const filter_forms = $('#kmcfmf_contact_form_7_filter_forms');
+            /*const filter_type = $('#kmcfmf_contact_form_7_filter_type');
+            filter_type.change((e) => {
+                e.preventDefault();
+                console.log(e.target.value)
+                if (e.target.value != '') {
+                    filter_forms.show();
 
+                }
+            })
+            filter_forms.hide();
+*/
+            filter_forms.selectize({
+                delimiter: ',',
+                persist: true,
+                options: <?php echo json_encode( $contact_forms )?>,
+            });
         })
     </script>
 <?php endif;